Debugging

+++ src/main/kotlin/com/example/dbsdevapp/repo/OrderRepo.kt
@@ -50,7 +50,7 @@ class OrderRepo(
 
     fun countAll() = SimpleJdbcCall(template)
         .withProcedureName("countOrders")
-        .execute()["count"]
+        .execute()["count"] as Int
 
     fun assignEmployeesToOrder(
         orderId: Int,
